Remove rawDependencyMapper and change dependencies name
